+++ SendBirdSDK.framework/Headers/SBDScheduledUserMessage.h
@@ -11,7 +11,7 @@
 #import "SBDSender.h"
 #import "SBDTypes.h"
 
-NS_ASSUME_NONNULL_BEGIN
+@class SBDMessageMetaArray;
 
 /**
  Represents a scheduled user message.
@@ -106,8 +106,17 @@ NS_ASSUME_NONNULL_BEGIN
 
 /**
  Meta array.
+ 
+ @deprecated 3.0.148
  */
-@property (nonatomic, nonnull, readonly, getter=getAllMetaArray) NSDictionary<NSString *, NSArray<NSString *> *> *metaArray;
+@property (nonatomic, nonnull, readonly, getter=getAllMetaArray) NSDictionary<NSString *, NSArray<NSString *> *> *metaArray DEPRECATED_MSG_ATTRIBUTE("Use -metaArrays");
+
+/**
+ Gets an array of meta arrays sorted by chronological order.
+ 
+ @since 3.0.148
+ */
+@property (strong, nonatomic, readonly, nonnull) NSArray<SBDMessageMetaArray *> *metaArrays;
 
 /**
  Checks the channel type is open channel or not.
@@ -128,9 +137,23 @@ NS_ASSUME_NONNULL_BEGIN
  
  @param keys Keys of the meta array.
  @return Meta array of the keys.
+ @see -metaArraysWithKeys:
+ @deprecated 3.0.148
  */
-- (nonnull NSDictionary<NSString *, NSArray<NSString *> *> *)getMetaArrayWithKeys:(NSArray<NSString *> * _Nonnull)keys;
+- (nonnull NSDictionary<NSString *, NSArray<NSString *> *> *)getMetaArrayWithKeys:(NSArray<NSString *> * _Nonnull)keys
+DEPRECATED_MSG_ATTRIBUTE("Use -metaArraysWithKeys:");
+
+/**
+ Returns an array of meta arrays for keys.
+ 
+ @param keys Keys of the meta array.
+ @return An Array of meta array with keys.
+ @code
+ NSArray<NSString *> *keys = meta_array_keys;
+ NSArray<SBDMessageMetaArray *> *subMetaArrays = [message metaArraysWithKeys:keys];
+ @endcode
+ @since 3.0.148
+ */
+- (nonnull NSArray<SBDMessageMetaArray *> *)metaArraysWithKeys:(nonnull NSArray<NSString *> *)keys;
 
 @end
-
-NS_ASSUME_NONNULL_END